Refactor: 모달창 합침 (공통 모달 + 비밀번호 모달) / 로그인, 회원가입 인터페이스 적용
