+++ src/interfaces/modal.ts
@@ -1,15 +1,10 @@
-import { ROUTES } from '@/constants/routes';
-
 export interface IModal {
   isOpen: boolean;
   text: string;
-  onClickOk: Function;
+  onClickOk: any;
   onClickCancel?: Function;
   okText?: string;
   cancelText?: string;
-}
-
-export interface PasswordModal {
-  isOpen: boolean;
-  linkURL: string;
+  isPassword?: boolean;
+  linkURL?: string;
 }
