+++ src/interfaces/store.ts
@@ -1,11 +1,6 @@
-import { IModal, PasswordModal } from './modal';
+import { IModal } from './modal';
 
 export interface IStore {
   loading: Boolean;
   modal: IModal;
-}
-
-export interface PasswordModalStore {
-  loading: Boolean;
-  passwordModal: PasswordModal;
 }
